Split off and skip precompilation tests on <1.11
